refactor: DietRepository 인터페이스 분리 및 아키텍처 개선
