+++ domain/src/main/java/com/cases/carefull/domain/usecase/bmr/GetSavedBmrUseCase.kt
@@ -1,13 +1,13 @@
 package com.cases.carefull.domain.usecase.bmr
 
 import com.cases.carefull.domain.model.diet.Bmr
-import com.cases.carefull.domain.repository.DietRepository
+import com.cases.carefull.domain.repository.diet.BodyStatsRepository
 import kotlinx.coroutines.flow.Flow
 
 class GetSavedBmrUseCase(
-	private val dietRepository: DietRepository
+	private val bodyStatsRepository: BodyStatsRepository
 ) {
 	operator fun invoke(userId: String): Flow<Bmr?> {
-		return dietRepository.getMyBmr(userId)
+		return bodyStatsRepository.getMyBmr(userId)
 	}
 }
