+++ domain/src/main/java/com/cases/carefull/domain/usecase/bmr/SaveBmrUseCase.kt
@@ -1,12 +1,12 @@
 package com.cases.carefull.domain.usecase.bmr
 
 import com.cases.carefull.domain.model.diet.Bmr
-import com.cases.carefull.domain.repository.DietRepository
+import com.cases.carefull.domain.repository.diet.BodyStatsRepository
 
 class SaveBmrUseCase(
-	private val dietRepository: DietRepository
+	private val bodyStatsRepository: BodyStatsRepository
 ) {
 	suspend operator fun invoke(bmrToSave: Bmr) {
-		dietRepository.insertBmr(bmrToSave)
+		bodyStatsRepository.updateMyBmr(bmrToSave)
 	}
 }
