+++ features/carefullmainui/src/main/java/com/cases/carefull/features/carefullmainui/home/HomeViewModel.kt
@@ -7,8 +7,8 @@ import com.cases.carefull.domain.model.exercise.ExerciseCollection
 import com.cases.carefull.domain.model.exercise.ExerciseRecordForDate
 import com.cases.carefull.domain.model.exercise.ExerciseType
 import com.cases.carefull.domain.repository.CalendarRepository
-import com.cases.carefull.domain.repository.DietRepository
 import com.cases.carefull.domain.repository.ExerciseRepository
+import com.cases.carefull.domain.repository.diet.DietRecordRepository
 import com.cases.carefull.domain.usecase.bmr.GetSavedBmrUseCase
 import com.cases.carefull.domain.util.DataResourceResult
 import com.cases.carefull.features.carefullmainui.home.HomeUiState.Companion.START_PAGE
@@ -35,7 +35,7 @@ import java.time.temporal.TemporalAdjusters
 class HomeViewModel @Inject constructor(
 	private val calendarRepository: CalendarRepository,
 	private val exerciseRepository: ExerciseRepository,
-	private val dietRepository: DietRepository,
+	private val dietRecordRepository: DietRecordRepository,
 	private val getSavedBmrUseCase: GetSavedBmrUseCase,
 ) : ViewModel() {
 	private val _uiState = MutableStateFlow(HomeUiState())
@@ -65,7 +65,7 @@ class HomeViewModel @Inject constructor(
 			combine(
 				exerciseRepository.getExerciseStatFlow("test"),
 				exerciseRepository.getCompletedDailyExerciseDatesFlow("test"),
-				dietRepository.getAllMeal(),
+				dietRecordRepository.getAllMeal(),
 				_uiState.map { it.viewType to it.selectedDate }.distinctUntilChanged()
 			) { exerciseStats, completedDates, dietResult, (viewType, selectedDate) ->
